Fix the table header field labels (see #2).
Description
-----------

In case that the language variable does not consist of an array, the field labels in the table header were not output correctly. This PR fixes the problem.

Furthermore, the CSS class  was not shown because it was commented out in the source code.

(applies to Contao 3.5 as well)

Commits
-------

dabf26fe Fix the table header field labels
9990332c Move the default value out of the else condition